sass works a little different 'round here
jekyll uses a _sass folder and underbar '_' prefix naming convention for sass partials. It also appears to use different namespace rules than however rails is set up.

this adjusts the scss to use the naming convention, splits out the partials from the files that will end up as css, and adds a missing import in order to get things to compile. It also comments out both 'global-reset' and 'compass', both used only on the error page. I'm uncertain if error will still be a reachable page from a static site, so I'm going to punt on this for now.

todo: figure out what to do about compass and global-reset