move assets over, get index working
this moves over the assets from the main site, puts them in a useful place, and does some mass file renames so the paths work out. It also makes some changes to the templating for the monitor -- the include statement can pass variables but those partial templates it loads cannot see page or site variables, only what are passed in. This will set the latest version and publish date site wide in _config.yml and propogate it through the appropriate spots in the website.

only the index is done, but this is a good proof of concept and logical stopping point for today